feat(search): change discover query to Rick Riordan
Updates the default search term for the "discover" feature on the
Search Page from "Classic Literature" to "Rick Riordan".

This change provides a more focused and engaging set of initial
results for users exploring the platform's search capabilities.

The corresponding unit tests have been updated to align with this
new default search behavior.